Created Main Menu and learned how to transition scenes in Godot
For this version, I just wanted to quickly go through how to set up scenes in Godot and load them. Godot has a very different way of managing levels and sprites compared to Unity, so I was thrown off a bit. I just created a main menu to try out loading and unloading scenes and played around with how to set-up the project and the GUI. I also increased the default size of the game window and made the GUI responsive to changes in window size. The art styles in this version are just placeholders I found online and in itch.io. The current load scenes are hard coded, something I am not too happy about. For this time, I just wanted to get a working version of the scene loading of one room. For my next commit, I will probably work on the inventory system just to figure out what that would look like in Godot. Afterward, I think that the next steps involve sitting down with good'ol pen and paper to figure out the game management system. I can already foresee that without a very clear plan of how to manage player, room, and level states this project will most likely be very hardcoded (and possibly finicky). I think taking a step back to just figure out the logic of it all will be good in the long run of this project. I also want to learn how to do smooth scene transitions (loading screens most likely)