Added interactable objects and inventory
I added the scripts to move objects in the point clicker style: press the mouse, and you can move it to the cursor's position until you press the mouse again. The other thing I worked on was making an inventory with items and being able to drag those items into different slots. Today was mainly concentrating on making this very basic mechanics work. The most important thing this week is to get the inventory to keep an accurate track of the items the player collects, especially as they move around the house's walls. Again, this will be easier once I sit down to really map out the UX and all of the components I need for the game manager. Oh, I also added a moving sprite because I want to be able to trigger a change in animation after moving an object to its same position. Collisions are not working very well, so I will fix them in my next work session.