Added UI folder and started restructuring the scene manager
I sketched out the logic for the scene manager of the game. Basically, I figured out the components needed so that the components are easy to reuse and allow us to move through scenes more efficiently. For this particular session, I created the UI elements we will need the most: the inventory and the arrows that let you move through the rooms. The will be one big UI component for each house so that the inventory for that house is saved throughout the player's traversal through the room. I've completed the toggle button for the inventory and added the room transition arrows. I know it doesn't look like much for now, but I think that making some components from zero rather than refactoring the ones I made last week will be easier with my new manager.

My current approach to the scene manager is based on a pseudocode sketch I did and a wireframe I made on Figma. For more information on this, go to the MDM folder and check out my (Thais) post.

I am also trying to plan ahead for the player-manager. I am still unsure of how to make the player-manager in Godot, but once I figure out the scene and the inventory, I believe the player-manager will be easier to do.