Added interaction animations for sewing room and entrance
Added assets and interactions to the sewing room and the entrance.

For the entrance, you can open and close the window. You can also take the piece of wood and barricade the door.

For the sewing machine room, the sewing machine pedal will have a jammed animation. A variable checks whether or not the nail jamming the sewing machine is there or not. If it is not there, then the sewing machine can work correctly.