Tried art layout and filters for the neighborhood scene
We (Thais, Savannah, and Jack) met and tested the look of the game. Savannah gave Thais the assets and we tested them in a scene called test.tscn. We tested 3 filters that will give a coherent look to the aesthetics of the game. We will test one more filter later today which will be more brush-stroke-like.

Jack gave feedback on whether the assets followed Chinese traditional furniture and Savannah adjusted accordingly. Through this back and forth, we are making sure that we have assets that authentically takes inspiration from Chinese culture while keeping in mind how the furniture in the other house of the game will change due to a "Westernization" of the decor.