Add files via upload
Just a fella. well, two fellas

I'm still trying to steer away from "this dude is clearly japanese" , as far as I'm aware we are still trying to keep the nationalities and whatnot rather vague and fluid, more fantastical than historical. i am not brave enough to slap elf ears on this dude, but if it were only my game, he'd 100% be an elf. 

i have a meeting in 7 hours YIPEEE