Add files via upload
"Damn! That's an ugly-ass baby. Eyes cocked like a pistol."

I played around with filters for the toyol, i think it'd be best to cycle through these with a loop. this piece is what made me realize i desperately wish i had used a limited color palette for the art assets, but alas, i am but a man. 

the kid was nice to draw. she looks creepy-cute, in a way that i 100% expect her to have a twin and stand at the end of long hallways, just to fuck with people. i kept her color palette the same as the toyol so hopefully the audience can put two n two together. there's a saying about not assuming your audience is dumb, and i think i can comfortably do that (as in, i think my peers are intelligen- ok this is an actual ramble and i have too much work to do. THE ART IS COMPLETED, COMMIT CHANGES