Playable version of opening house (3) with art assets
This is a playable version of the house that opens the game (house 3). States of objects and rooms work for the most part. A few state elements are needed, but all the interactions are done and accounted for.

Today Savannah and Thais(me) set down to import all the assets of the first house (house 3). We spent much of the day porting the assets and making sure that the game did not break. Jack, on the other hand, is getting sound assets to use in the game. He was also tasked with making the neighborhood scene that will allow the user to navigate between houses. We will be working on combining his stuff during the weekend.

During the past weekend, I (Thais) refactored some of the scenes to use the Popochiu addon for Godot. Using the Popochiu plugin allowed me to have templates of nodes and code such as rooms, props, and hotspots for clicking. This helped me recycle code more easily and the templates allowed for an easier way of changing the sprites. Basically, everything in the object stays the same except for the sprite sheet. Once the final sprite is imported, I can quickly adjust things like colliders, ratios and positioning to better fit the new sprite. Very little tweaking of the code is needed.

I am very happy to have found the Popochiu plugin. It has really made my life easier and its comes with many great built-in assets. It was also created by a Latin American developer with the intention of making it easier to jump right into the state logic of point-clicker games rather than focusing on getting the basics such as clicking and adding to inventory.