Implementing state conditions for unlocking house 5
Currently working on implementing the saving, loading, and checking for the state conditions that allow the player to unlock house 5 (the second house in the playthrough). This is relatively easy in Godot. I am using a .gd script called Player state and each room has a state.gd script as well that allows me to save data on room exit and then load it on room entered. So far, the states for the kitchen, sewing room, entrance, and altar work. Just missing the bedroom and then triggering a cutscene for when the room is completed as well as unlocking house 5 for the player to exit house 3 (first house in playthrough)