Added Main Menu -- 100% finished with first house interactions and conditions
Savannah created some beautiful art for the main menu. I implemented it and set up all of the buttons for the main menu.

I finalized all of the conditions to start the game. This build so far could be the last one if we wanted to demo only the openning house. Will continue to work on sound