Added Main Menu and started implementing Jack's version of neighborhood transition scene
Changed the layout of the main menu for aesthetic purposes. By popular request, I also added our roles in the menu as a credits section.

For the neighborhood scene (lobby scene), I am working on implementing Jack's layout (previously in the MDM file). I am adding the scene transition elements as well as a parallax effect for when the player moves around the scene.