Final version for class presentation
This is a working version of the play-through of the game. The first house is completely finished with case states and represents a working version of how the game would work/feel.

The first house ends with a cutscene that prompts the user to go back to the main menu or continue to the neighborhood scene, where they can access the other houses. The other two houses are currently under development. They mainly show the key aspects of how they would differ from one another, as well as the flow between houses. Some interactions do affect the state of the first and second houses. However, the state machine for all three houses is not complete.

Very proud of this game. Technically speaking, if we had one more week, I could complete the whole 5 houses because of the way I structured the workflow. I was able to place in houses 1 and 5 fairly quickly, and I got a working inventory item pickup as well. Most of the interactions are very similar to the first house so the programming logic is very much the same. The issue is keeping track of the states amongst the houses (having one house's state affect the other).

I honestly also like the MDM workflow. I think that it is very helpful to have executable versions of each iteration of the game. The only thing I have against it is that I've committed to GitHub less often than I would because I feel a bit of pressure of writing something meaningful in each commit.

For this past week, I tried to take that pressure off and I think it's resulted in a more healthy relationship with how I can use MDM to adjust to my needs. I think this is a method that I will be revisiting for other projects.